+++ src/main/java/pl/pjatk/jaz_s32698_nbp/dto/Rate.java
@@ -1,7 +1,9 @@
 package pl.pjatk.jaz_s32698_nbp.dto;
 
+import java.math.BigDecimal;
+
 public class Rate {
-    private double mid;
-    public double getMid() { return mid; }
-    public void setMid(double mid) { this.mid = mid; }
+    private BigDecimal mid;
+    public BigDecimal getMid() { return mid; }
+    public void setMid(BigDecimal mid) { this.mid = mid; }
 }
